main: extract stop_signal into its own header
To make it reusable in tools as well.
